+++ R/ctrGenerateQueries.R
@@ -29,7 +29,8 @@
 #' @param completedAfter String that can be interpreted as date (does not work
 #' with EUCTR)
 #' @param onlyWithResults Logical
-#' @param registers Vector of register names, default all registers
+#' @param countries Vector of country names, two- or three-letter
+#' ISO 3166 codes
 #'
 #' @returns Named vector of URLs for finding trials in the registers and as
 #' input to functions \link{ctrLoadQueryIntoDb} and
@@ -63,7 +64,8 @@
 #'    "basket OR platform OR umbrella OR master protocol OR ",
 #'    "multiarm OR multistage OR subprotocol OR substudy OR ",
 #'    "multi-arm OR multi-stage OR sub-protocol OR sub-study"),
-#'  startAfter = "2010-01-01")
+#'  startAfter = "2010-01-01",
+#'  countries = c("DE", "US", "United Kingdom"))
 #'
 #' # open queries in register web interface
 #' sapply(urls, ctrOpenSearchPagesInBrowser)
@@ -80,16 +82,23 @@ ctrGenerateQueries <- function(
     completedBefore = NULL,
     completedAfter = NULL,
     onlyWithResults = FALSE,
-    registers = c("EUCTR", "ISRCTN", "CTIS", "CTGOV2")
+    countries = NULL
 ) {
 
   # check
-  stopifnot(all(registers %in% registerList[registerList != "CTGOV"]))
+  stopifnot(all(countries %in% unlist(countryTable)))
 
+  # map to row in table
+  countryIndex <- seq_len(nrow(countryTable))[apply(
+    countryTable, 1, function(r)
+      length(intersect(unlist(r), countries)) > 0L)]
 
   #### start ####
   urls <- c(
     "CTGOV2" = "https://clinicaltrials.gov/search?",
+    # https://clinicaltrials.gov/find-studies/constructing-complex-search-queries
+    # https://clinicaltrials.gov/data-api/about-api/search-areas
+    "CTGOV2expert" = "https://clinicaltrials.gov/expert-search?term=",
     "EUCTR" = "https://www.clinicaltrialsregister.eu/ctr-search/search?query=",
     "ISRCTN" = "https://www.isrctn.com/search?",
     "CTIS" = "https://euclinicaltrials.eu/ctis-public/search#searchCriteria={"
@@ -112,6 +121,8 @@ ctrGenerateQueries <- function(
 
     searchPhraseC <- gsub("( OR | AND )", ", ", searchPhrase)
 
+    searchPhraseD <- stringi::stri_extract_all_regex(searchPhrase, "( OR | AND )")[[1]][1]
+
     if (grepl(" AND ", searchPhrase) && grepl(" OR ", searchPhrase)) {
       warning(
         'Cannot use both "AND" and "OR" with CTIS. ',
@@ -127,6 +138,13 @@ ctrGenerateQueries <- function(
     # "basket" OR "platform" OR "umbrella" OR "master protocol" OR "multiarm" OR "multistage" OR "subprotocol" OR "substudy" OR "multi-arm" OR "multi-stage" OR "sub-protocol" OR "sub-study"
     # title search 2372
     # https://clinicaltrials.gov/search?titles=%22basket%22%20OR%20%22platform%22%20OR%20%22umbrella%22%20OR%20%22master%20protocol%22%20OR%20%22multiarm%22%20OR%20%22multistage%22%20OR%20%22subprotocol%22%20OR%20%22substudy%22%20OR%20%22multi-arm%22%20OR%20%22multi-stage%22%20OR%20%22sub-protocol%22%20OR%20%22sub-study%22
+
+    # ctgov2expert
+    urls["CTGOV2expert"] <- paste0(
+      urls["CTGOV2expert"], " AND (",
+      paste0('AREA[TitleSearch]"', searchPhraseA, '"', collapse = searchPhraseD),
+      ") "
+    )
 
     # ctis
     urls["CTIS"] <- paste0(
@@ -170,6 +188,9 @@ ctrGenerateQueries <- function(
     urls["CTGOV2"] <- paste0(
       urls["CTGOV2"], "cond=", condition)
 
+    urls["CTGOV2expert"] <- paste0(
+      urls["CTGOV2expert"], ' AND AREA[ConditionSearch]"', condition, '" ')
+
     urls["CTIS"] <- paste0(
       urls["CTIS"], '"medicalCondition":"', condition, '",')
 
@@ -188,6 +209,10 @@ ctrGenerateQueries <- function(
     urls["CTGOV2"] <- paste0(
       urls["CTGOV2"], "&intr=", intervention)
 
+    # ctgov2expert
+    urls["CTGOV2expert"] <- paste0(
+      urls["CTGOV2expert"], ' AND AREA[InterventionSearch]"', intervention, '" ')
+
     urls["CTIS"] <- paste0(
       urls["CTIS"], '"containAll":"', intervention, '",')
 
@@ -205,7 +230,7 @@ ctrGenerateQueries <- function(
 
     stopifnot(is.atomic(phase) && length(phase) == 1L)
 
-    # see also f.trialPhase
+    # see also f_trialPhase.R
 
     urls["CTGOV2"] <- paste0(
       urls["CTGOV2"], "&aggFilters=phase:", c(
@@ -220,6 +245,21 @@ ctrGenerateQueries <- function(
         "phase 1+2+3+4" = "0 1 2 3 4"
       )[phase]
     )
+
+    # https://clinicaltrials.gov/data-api/about-api/study-data-structure#enum-Phase
+    urls["CTGOV2expert"] <- paste0(
+      urls["CTGOV2expert"], ' AND (', c(
+        "phase 1" = 'AREA[Phase]"EARLY_PHASE1" OR AREA[Phase]"PHASE1"',
+        "phase 1+2" = 'AREA[Phase]"EARLY_PHASE1" OR AREA[Phase]"PHASE1" OR AREA[Phase]"PHASE2"',
+        "phase 2" = 'AREA[Phase]"PHASE2"',
+        "phase 2+3" = 'AREA[Phase]"PHASE2" OR AREA[Phase]"PHASE3"',
+        "phase 3" = 'AREA[Phase]"PHASE3"',
+        "phase 3+4" = 'AREA[Phase]"PHASE3" OR AREA[Phase]"PHASE4"',
+        "phase 1+2+3" = 'AREA[Phase]"EARLY_PHASE1" OR AREA[Phase]"PHASE1" OR AREA[Phase]"PHASE2" OR AREA[Phase]"PHASE3"',
+        "phase 4" = 'AREA[Phase]"PHASE4"',
+        "phase 1+2+3+4" = 'AREA[Phase]"EARLY_PHASE1" OR AREA[Phase]"PHASE1" OR AREA[Phase]"PHASE2" OR AREA[Phase]"PHASE3" OR AREA[Phase]"PHASE4"'
+      )[phase],
+      ') ')
 
     urls["CTIS"] <- paste0(
       urls["CTIS"], '"trialPhaseCode":[', c(
@@ -281,6 +321,18 @@ ctrGenerateQueries <- function(
       )[population]
     )
 
+    # https://clinicaltrials.gov/data-api/about-api/study-data-structure#enum-StandardAge
+    urls["CTGOV2expert"] <- paste0(
+      urls["CTGOV2expert"], ' AND (', c(
+        "P" = 'AREA[StdAge]"CHILD"',
+        "A" = 'AREA[StdAge]"ADULT"',
+        "E" = 'AREA[StdAge]"OLDER_ADULT"',
+        "P+A" = 'AREA[StdAge]"CHILD" OR AREA[StdAge]"ADULT"',
+        "A+E" = 'AREA[StdAge]"ADULT" OR AREA[StdAge]"OLDER_ADULT"',
+        "P+A+E" = 'AREA[StdAge]"CHILD" OR AREA[StdAge]"ADULT" OR AREA[StdAge]"OLDER_ADULT"'
+      )[phase],
+      ') ')
+
     urls["CTIS"] <- paste0(
       urls["CTIS"], '"ageGroupCode":[', c(
         # 1 = in utero
@@ -333,6 +385,23 @@ ctrGenerateQueries <- function(
       )[recruitment]
     )
 
+    # https://clinicaltrials.gov/data-api/about-api/study-data-structure#enum-RecruitmentStatus
+    urls["CTGOV2expert"] <- paste0(
+      urls["CTGOV2expert"], ' AND (', c(
+        "ongoing" = paste0(
+          'AREA[LocationStatus]"ACTIVE_NOT_RECRUITING" OR ',
+          'AREA[LocationStatus]"ENROLLING_BY_INVITATION" OR ',
+          'AREA[LocationStatus]"RECRUITING"'),
+        "completed" = 'AREA[LocationStatus]"COMPLETED"',
+        "other" = paste0(
+          'AREA[LocationStatus]"NOT_YET_RECRUITING" OR ',
+          'AREA[LocationStatus]"SUSPENDED" OR ',
+          'AREA[LocationStatus]"TERMINATED" OR ',
+          'AREA[LocationStatus]"WITHDRAWN" OR ',
+          'AREA[LocationStatus]"AVAILABLE"')
+      )[recruitment],
+      ') ')
+
     urls["CTIS"] <- paste0(
       urls["CTIS"], '"status":[', c(
         "ongoing" = "2,3,4,6,7",
@@ -366,6 +435,10 @@ ctrGenerateQueries <- function(
     urls["CTGOV2"] <- paste0(
       urls["CTGOV2"], "&start=", startAfter, "_")
 
+    urls["CTGOV2expert"] <- paste0(
+      urls["CTGOV2expert"], " AND AREA[StartDate]RANGE[",
+      startAfter, ",MAX] ")
+
     urls["CTIS"] <- paste0(
       urls["CTIS"], '"eeaStartDateFrom":"', startAfter, '",')
 
@@ -388,6 +461,11 @@ ctrGenerateQueries <- function(
       sub("(&start=[-0-9]+_)", paste0("\\1", startBefore), urls["CTGOV2"]),
       paste0(urls["CTGOV2"], "&start=_", startBefore))
 
+    urls["CTGOV2expert"] <- ifelse(
+      grepl("StartDate", urls["CTGOV2expert"]),
+      sub("(AREA\\[StartDate\\]RANGE\\[[-0-9]+,)MAX", paste0("\\1", startBefore), urls["CTGOV2expert"]),
+      paste0(urls["CTGOV2expert"], " AREA[StartDate]RANGE[MAX,",startBefore, "] "))
+
     urls["CTIS"] <- paste0(
       urls["CTIS"], '"eeaStartDateTo":"', startBefore, '",')
 
@@ -408,6 +486,10 @@ ctrGenerateQueries <- function(
     urls["CTGOV2"] <- paste0(
       urls["CTGOV2"], "&primComp=", completedAfter, "_")
 
+    urls["CTGOV2expert"] <- paste0(
+      urls["CTGOV2expert"], " AND AREA[CompletionDate]RANGE[",
+      completedAfter, ",MAX] ")
+
     urls["CTIS"] <- paste0(
       urls["CTIS"], '"eeaEndDateFrom":"', completedAfter, '",')
 
@@ -427,6 +509,11 @@ ctrGenerateQueries <- function(
       sub("(&primComp=[-0-9]+_)", paste0("\\1", completedBefore), urls["CTGOV2"]),
       paste0(urls["CTGOV2"], "&primComp=_", completedBefore))
 
+    urls["CTGOV2expert"] <- ifelse(
+      grepl("CompletionDate", urls["CTGOV2expert"]),
+      sub("(AREA\\[CompletionDate\\]RANGE\\[[-0-9]+,)MAX", paste0("\\1", completedBefore), urls["CTGOV2expert"]),
+      paste0(urls["CTGOV2expert"], " AREA[CompletionDate]RANGE[MAX,",completedBefore, "] "))
+
     urls["CTIS"] <- paste0(
       urls["CTIS"], '"eeaEndDateTo":"', completedBefore, '",')
 
@@ -442,6 +529,9 @@ ctrGenerateQueries <- function(
     urls["CTGOV2"] <- paste0(
       urls["CTGOV2"], "&aggFilters=results:with")
 
+    urls["CTGOV2expert"] <- paste0(
+      urls["CTGOV2expert"], "AND (NOT AREA[ResultsFirstPostDate]MISSING) ")
+
     urls["CTIS"] <- paste0(
       urls["CTIS"], '"hasClinicalStudyReport":true,')
 
@@ -453,10 +543,58 @@ ctrGenerateQueries <- function(
 
   }
 
-  # finalising CTIS
-  urls["CTIS"] <- paste0(sub(",$", "", urls["CTIS"]), '}')
 
-  # finalising CTGOV2
+  #### countries ####
+  if (!is.null(countries)) {
+
+    urls["CTGOV2expert"] <- paste0(
+      urls["CTGOV2expert"], " AND (",
+      paste0(
+        'AREA[LocationCountry]"',
+        countryTable[countryIndex, 2],
+        collapse = '" OR '
+      ), '") '
+    )
+
+    if (length(countryIndex) == 1L) {
+
+      urls["CTGOV2"] <- paste0(
+        urls["CTGOV2"],
+        "&country=", countryTable[countryIndex, 2][1]
+      )
+
+    } else {
+
+      urls["CTGOV2"] <- urls["CTGOV2expert"]
+
+    }
+
+    urls["CTIS"] <- paste0(
+      urls["CTIS"], '"msc":[',
+      paste0(countryTable[countryIndex, 1], collapse = ","),
+      '],')
+
+    tmpC <- countryTable[countryIndex, 3]
+    urls["EUCTR"] <- paste0(
+      urls["EUCTR"],
+      paste0(
+        "&country=",
+        c(tolower(tmpC[tmpC %in% countriesEUCTR]),
+          if(!all(tmpC %in% countriesEUCTR)) "3rd"),
+        collapse = "")
+    )
+
+    urls["ISRCTN"] <- paste0(
+      urls["ISRCTN"],
+      paste0("&filters=recruitmentCountry:", countryTable[countryIndex, 2], collapse = "")
+    )
+
+  }
+
+
+  #### finalising ####
+
+  # CTGOV2
   ctgov2Filter <- stringi::stri_extract_all_regex(
     urls["CTGOV2"], "[,&]aggFilters=[^,&]+")[[1]]
   if (!all(is.na(ctgov2Filter))) urls["CTGOV2"] <- paste0(
@@ -465,7 +603,14 @@ ctrGenerateQueries <- function(
     "&aggFilters=",
     paste0(sub("&aggFilters=", "", ctgov2Filter), collapse = ","))
 
-  # finalising ISRCTN
+  # CTGOVexpert
+  urls["CTGOV2"] <- sub("search[?]term= AND ", "search?term=", urls["CTGOV2"])
+  urls["CTGOV2expert"] <- sub("search[?]term= AND ", "search?term=", urls["CTGOV2expert"])
+
+  # CTIS
+  urls["CTIS"] <- paste0(sub(",$", "", urls["CTIS"]), '}')
+
+  # ISRCTN
   isrctnFilter <- stringi::stri_extract_all_regex(
     urls["ISRCTN"], "[,&]filters=[^,&]+")[[1]]
   if (!all(is.na(isrctnFilter))) urls["ISRCTN"] <- paste0(
@@ -476,9 +621,8 @@ ctrGenerateQueries <- function(
   if (!grepl("q=", urls["ISRCTN"])) urls["ISRCTN"] <-
     paste0(urls["ISRCTN"], "&q=")
 
-  # select but put CTIS last so that it would open
-  # last and run script to retrieve the results
-  urls <- urls[registerList[registerList %in% registers]]
+  # put CTIS last so that it would open
+  urls <- urls[c("EUCTR", "ISRCTN", "CTGOV2", "CTGOV2expert", "CTIS")]
 
   # named vector
   return(urls)