+++ inst/tinytest/test_ctrdata_function_ctrgeneratequeries.R
@@ -19,13 +19,14 @@ tf <- function() {
     onlyWithResults = TRUE)
 
   # test
-  expect_length(urls, 4)
+  expect_length(urls, 5L)
 
   # test
   expect_equivalent(c(
     "https://www.clinicaltrialsregister.eu/ctr-search/search?query=cancer AND antibody&dateFrom=2000-01-01&dateTo=2030-01-01&resultsstatus=trials-with-results",
-    "https://clinicaltrials.gov/search?cond=cancer&intr=antibody&start=2000-01-01_2030-01-01&primComp=2000-01-01_2030-01-01&aggFilters=results:with",
     "https://www.isrctn.com/search?&filters=condition:cancer,intervention:antibody,GT+overallStartDate:2000-01-01,LE+overallStartDate:2030-01-01,GT+overallEndDate:2000-01-01,LE+overallEndDate:2030-01-01,results:withResults&q=",
+    "https://clinicaltrials.gov/search?cond=cancer&intr=antibody&start=2000-01-01_2030-01-01&primComp=2000-01-01_2030-01-01&aggFilters=results:with",
+    "https://clinicaltrials.gov/expert-search?term=AREA[ConditionSearch]\"cancer\"  AND AREA[InterventionSearch]\"antibody\"  AND AREA[StartDate]RANGE[2000-01-01,2030-01-01]  AND AREA[CompletionDate]RANGE[2000-01-01,2030-01-01] AND (NOT AREA[ResultsFirstPostDate]MISSING) ",
     "https://euclinicaltrials.eu/ctis-public/search#searchCriteria={\"medicalCondition\":\"cancer\",\"containAll\":\"antibody\",\"eeaStartDateFrom\":\"2000-01-01\",\"eeaStartDateTo\":\"2030-01-01\",\"eeaEndDateFrom\":\"2000-01-01\",\"eeaEndDateTo\":\"2030-01-01\",\"hasClinicalStudyReport\":true}"
   ), urls)
 
@@ -36,13 +37,14 @@ tf <- function() {
     startBefore = "2030-01-01")
 
   # test
-  expect_length(urls, 4)
+  expect_length(urls, 5L)
 
   # test
   expect_equivalent(c(
     "https://www.clinicaltrialsregister.eu/ctr-search/search?query=bispecific&dateTo=2030-01-01",
-    "https://clinicaltrials.gov/search?&intr=bispecific&start=_2030-01-01",
     "https://www.isrctn.com/search?&filters=intervention:bispecific,LE+overallStartDate:2030-01-01&q=",
+    "https://clinicaltrials.gov/search?&intr=bispecific&start=_2030-01-01",
+    "https://clinicaltrials.gov/expert-search?term=AREA[InterventionSearch]\"bispecific\"  AREA[StartDate]RANGE[MAX,2030-01-01] ",
     "https://euclinicaltrials.eu/ctis-public/search#searchCriteria={\"containAll\":\"bispecific\",\"eeaStartDateTo\":\"2030-01-01\"}"
   ), urls)
 
@@ -50,16 +52,18 @@ tf <- function() {
   # generate 3
   urls <- ctrGenerateQueries(
     condition = "cardiac failure",
-    completedAfter = "2000-01-01",
-    registers = c("CTGOV2", "EUCTR"))
+    completedAfter = "2000-01-01")
 
   # test
-  expect_length(urls, 2)
+  expect_length(urls, 5L)
 
   # test
   expect_equivalent(c(
     "https://www.clinicaltrialsregister.eu/ctr-search/search?query=cardiac failure",
-    "https://clinicaltrials.gov/search?cond=cardiac failure&primComp=2000-01-01_"
+    "https://www.isrctn.com/search?&filters=condition:cardiac failure,GT+overallEndDate:2000-01-01&q=",
+    "https://clinicaltrials.gov/search?cond=cardiac failure&primComp=2000-01-01_",
+    "https://clinicaltrials.gov/expert-search?term=AREA[ConditionSearch]\"cardiac failure\"  AND AREA[CompletionDate]RANGE[2000-01-01,MAX] ",
+    "https://euclinicaltrials.eu/ctis-public/search#searchCriteria={\"medicalCondition\":\"cardiac failure\",\"eeaEndDateFrom\":\"2000-01-01\"}"
   ), urls)
 
 
@@ -70,13 +74,14 @@ tf <- function() {
     startAfter = "2000-01-01")
 
   # test
-  expect_length(urls, 4)
+  expect_length(urls, 5L)
 
   # test
   expect_equivalent(c(
     "https://www.clinicaltrialsregister.eu/ctr-search/search?query=cancer&status=ongoing&status=trial-now-transitioned&status=suspended-by-ca&status=temporarily-halted&status=restarted&dateFrom=2000-01-01",
-    "https://clinicaltrials.gov/search?cond=cancer&start=2000-01-01_&aggFilters=status:act rec",
     "https://www.isrctn.com/search?&filters=condition:cancer,trialStatus:ongoing,GT+overallStartDate:2000-01-01&q=",
+    "https://clinicaltrials.gov/search?cond=cancer&start=2000-01-01_&aggFilters=status:act rec",
+    "https://clinicaltrials.gov/expert-search?term=AREA[ConditionSearch]\"cancer\"  AND (AREA[LocationStatus]\"ACTIVE_NOT_RECRUITING\" OR AREA[LocationStatus]\"ENROLLING_BY_INVITATION\" OR AREA[LocationStatus]\"RECRUITING\")  AND AREA[StartDate]RANGE[2000-01-01,MAX] ",
     "https://euclinicaltrials.eu/ctis-public/search#searchCriteria={\"medicalCondition\":\"cancer\",\"status\":[2,3,4,6,7],\"eeaStartDateFrom\":\"2000-01-01\"}"
   ), urls)
 
@@ -88,13 +93,14 @@ tf <- function() {
     startAfter = "2000-01-01")
 
   # test
-  expect_length(urls, 4)
+  expect_length(urls, 5L)
 
   # test
   expect_equivalent(c(
     "https://www.clinicaltrialsregister.eu/ctr-search/search?query=heart failure&phase=phase-two&phase=phase-three&dateFrom=2000-01-01",
-    "https://clinicaltrials.gov/search?cond=heart failure&start=2000-01-01_&aggFilters=phase:2 3",
     "https://www.isrctn.com/search?&filters=condition:heart failure,phase:Phase II/III,GT+overallStartDate:2000-01-01&q=",
+    "https://clinicaltrials.gov/search?cond=heart failure&start=2000-01-01_&aggFilters=phase:2 3",
+    "https://clinicaltrials.gov/expert-search?term=AREA[ConditionSearch]\"heart failure\"  AND (AREA[Phase]\"PHASE2\" OR AREA[Phase]\"PHASE3\")  AND AREA[StartDate]RANGE[2000-01-01,MAX] ",
     "https://euclinicaltrials.eu/ctis-public/search#searchCriteria={\"medicalCondition\":\"heart failure\",\"trialPhaseCode\":[10],\"eeaStartDateFrom\":\"2000-01-01\"}"
   ), urls)
 
@@ -104,18 +110,20 @@ tf <- function() {
     condition = "heart failure",
     phase = "phase 2+3",
     population = "P+A",
-    startAfter = "2000-01-01")
+    startAfter = "2000-01-01",
+    countries = c("Germany", "SE"))
 
   # test
-  expect_length(urls, 4)
+  expect_length(urls, 5L)
 
   # test
   expect_equivalent(c(
-    "https://www.clinicaltrialsregister.eu/ctr-search/search?query=heart failure&phase=phase-two&phase=phase-three&age=children&age=adolescent&age=infant-and-toddler&age=newborn&age=preterm-new-born-infants&age=under-18&age=adult&dateFrom=2000-01-01",
-    "https://clinicaltrials.gov/search?cond=heart failure&start=2000-01-01_&aggFilters=phase:2 3,ages:child adult",
-    "https://www.isrctn.com/search?&filters=condition:heart failure,phase:Phase II/IIINA,GT+overallStartDate:2000-01-01&q=",
-    "https://euclinicaltrials.eu/ctis-public/search#searchCriteria={\"medicalCondition\":\"heart failure\",\"trialPhaseCode\":[10],\"ageGroupCode\":[2,3],\"eeaStartDateFrom\":\"2000-01-01\"}"
-  ), urls)
+    "https://www.clinicaltrialsregister.eu/ctr-search/search?query=heart failure&phase=phase-two&phase=phase-three&age=children&age=adolescent&age=infant-and-toddler&age=newborn&age=preterm-new-born-infants&age=under-18&age=adult&dateFrom=2000-01-01&country=de&country=se",
+    "https://www.isrctn.com/search?&filters=condition:heart failure,phase:Phase II/IIINA,GT+overallStartDate:2000-01-01,recruitmentCountry:Germany,recruitmentCountry:Sweden&q=",
+    "https://clinicaltrials.gov/expert-search?term=AREA[ConditionSearch]\"heart failure\"  AND (AREA[Phase]\"PHASE2\" OR AREA[Phase]\"PHASE3\")  AND (NA)  AND AREA[StartDate]RANGE[2000-01-01,MAX]  AND (AREA[LocationCountry]\"Germany\" OR AREA[LocationCountry]\"Sweden\") ",
+    "https://clinicaltrials.gov/expert-search?term=AREA[ConditionSearch]\"heart failure\"  AND (AREA[Phase]\"PHASE2\" OR AREA[Phase]\"PHASE3\")  AND (NA)  AND AREA[StartDate]RANGE[2000-01-01,MAX]  AND (AREA[LocationCountry]\"Germany\" OR AREA[LocationCountry]\"Sweden\") ",
+    "https://euclinicaltrials.eu/ctis-public/search#searchCriteria={\"medicalCondition\":\"heart failure\",\"trialPhaseCode\":[10],\"ageGroupCode\":[2,3],\"eeaStartDateFrom\":\"2000-01-01\",\"msc\":[84,214]}"
+    ), urls)
 
 
   # generate 7
@@ -124,13 +132,14 @@ tf <- function() {
     recruitment = "completed")
 
   # test
-  expect_length(urls, 4)
+  expect_length(urls, 5L)
 
   # test
   expect_equivalent(c(
     "https://www.clinicaltrialsregister.eu/ctr-search/search?query=\"antibody\" AND \"covid\"&status=completed",
-    "https://clinicaltrials.gov/search?titles=\"antibody\" AND \"covid\"&aggFilters=status:com",
     "https://www.isrctn.com/search?q=\"antibody\" AND \"covid\"&filters=trialStatus:completed",
+    "https://clinicaltrials.gov/search?titles=\"antibody\" AND \"covid\"&aggFilters=status:com",
+    "https://clinicaltrials.gov/expert-search?term=(AREA[TitleSearch]\"antibody\" AND AREA[TitleSearch]\"covid\")  AND (AREA[LocationStatus]\"COMPLETED\") ",
     "https://euclinicaltrials.eu/ctis-public/search#searchCriteria={\"containAll\":\"antibody, covid\",\"status\":[5,8]}"
   ), urls)